requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
greenlet==3.1.1
psycopg2==2.9.9
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
sqlparse==0.5.1
typing_extensions==4.12.2
